moving leaderboard folder into pages
