--- frontend/app/src/leaderboard/index.ts
@@ -1,1 +1,0 @@
-export * from './store';
