+++ frontend/app/src/pages/leaderboard/store.ts
@@ -2,7 +2,7 @@
 import { orderBy } from 'lodash';
 import memo from 'memo-decorator';
 import { makeAutoObservable } from 'mobx';
-import api from '../api';
+import api from '../../api';
 
 export type LeaderItem = {
   owner_pubkey: string;
